+++ CHANGELOG.md
@@ -9,13 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - ForwardTriggerEvents component.
+- [UIFader] FadeInCoroutine/FadeOutCoroutine methods that return the Coroutine handle.
 
 ### Changed
-- DetectGround sets IsOnGround to false on disable
-- Send exit events on disable from ForwardTriggerEvents
+- DetectGround sets IsOnGround to false on disable.
+- Send exit events on disable from ForwardTriggerEvents.
+- [UIFader] FadeIn/FadeOut no longer return a Coroutine handle (allows use as UnityEvent listener).
 
 ### Fixed
-- [DetectGround] Update LastHit before sending OnGroundChanged event
+- [DetectGround] Update LastHit before sending OnGroundChanged event.
 
 ## [1.1.0] - 2023-02-08